backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.126.0",
    "langchain>=1.2.0",
    "langchain-groq>=1.1.1",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "uvicorn[standard]>=0.38.0",
]
